1005 1:N foreignkey

+++ crud/articles/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect
-from .models import Article
-from .forms import ArticleForm
+from .models import Article, Comment
+from .forms import ArticleForm, CommentForm
 # Create your views here.
 
 def index(request):
@@ -27,7 +27,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -40,16 +42,21 @@ def create(request):
 
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    comment_form = CommentForm()
+    comments = article.comment_set.all()
     context = {
         'article' : article,
+        'comment_form' : comment_form,
+        'comments' : comments,
     }
     return render(request, 'articles/detail.html', context)
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
     if request.method == "POST":
-        article.delete()
-        return redirect('articles:index')
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
     return redirect('articles:detail', article.pk)
 
 # def edit(request, pk):
@@ -67,15 +74,35 @@ def update(request, pk):
     # article.content = request.POST.get('content')
     # article.save()
     # return render(request, 'articles/create.html')
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'form' : form,
         'article' : article,
     }
     return render(request, 'articles/update.html', context)
+
+
+def comments_create(request, pk):
+    article = Article.objects.get(pk=pk)
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False)
+        comment.article = article
+        comment.user = request.user
+        comment.save()
+    return redirect('articles:detail', article.pk)
+
+def comments_delete(request, article_pk, comment_pk):
+    comment = Comment.objects.get(pk=comment_pk)
+    if request.user == comment.user:
+        comment.delete()
+    return redirect('articles:detail', article_pk)
